@@ -3,10 +3,10 @@
     "description": "Implementation of the specification pattern in PHP",
     "homepage": "https://github.com/RiskioFr/Specification",
     "require": {
-        "php": "^7.0"
+        "php": "^8.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.3"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
