@@ -1,16 +1,17 @@
 <?php
 namespace Riskio\SpecificationTest;
 
+use PHPUnit\Framework\TestCase;
 use Riskio\Specification\OrSpecification;
 use Riskio\SpecificationTest\Fixtures\NotSatisfiedSpecification;
 use Riskio\SpecificationTest\Fixtures\SatisfiedSpecification;
 
-class OrSpecificationTest extends \PHPUnit_Framework_TestCase
+class OrSpecificationTest extends TestCase
 {
     /**
      * @test
      */
-    public function isSatisfiedBy_GivenBothSpecificationsSatisfied_ShouldReturnTrue()
+    public function isSatisfiedBy_GivenBothSpecificationsSatisfied_ShouldReturnTrue(): void
     {
         $spec = new OrSpecification(
             new SatisfiedSpecification(),
@@ -25,7 +26,7 @@ class OrSpecificationTest extends \PHPUnit_Framework_TestCase
     /**
      * @test
      */
-    public function isSatisfiedBy_GivenOneSpecificationSatisfied_ShouldReturnTrue()
+    public function isSatisfiedBy_GivenOneSpecificationSatisfied_ShouldReturnTrue(): void
     {
         $spec = new OrSpecification(
             new SatisfiedSpecification(),
@@ -40,7 +41,7 @@ class OrSpecificationTest extends \PHPUnit_Framework_TestCase
     /**
      * @test
      */
-    public function isSatisfiedBy_GivenBothSpecificationsNotSatisfied_ShouldReturnFalse()
+    public function isSatisfiedBy_GivenBothSpecificationsNotSatisfied_ShouldReturnFalse(): void
     {
         $spec = new OrSpecification(
             new NotSatisfiedSpecification(),
